- make type of 'prog' u_long to be consistent

+++ trunk/lmbench3/src/lib_udp.h
@@ -5,7 +5,7 @@
 #include	<arpa/inet.h>
 
 int	udp_server(u_long prog, int rdwr);
-void	udp_done(int prog);
+void	udp_done(u_long prog);
 int	udp_connect(char *host, u_long prog, int rdwr);
 void	sock_optimize(int sock, int rdwr);
 int	sockport(int);
